src/main/java/io/github/bfeng/leetcode/PalindromeNumber.java
package io.github.bfeng.leetcode;

/**
 * Created by bfeng on 1/2/16.
 */
public class PalindromeNumber {
    public boolean isPalindrome(int x) {
        if (x < 0) return false;
        int left;
        int right;
        int div = 1;
        while (x / div >= 10) {
            div *= 10;
        }
        while (x != 0) {
            left = x / div;
            right = x % 10;
            if (left != right) return false;
            x = (x % div) / 10;
            div /= 100;
        }
        return true;
    }
}
